+++ packages/core/styles/src/gradients/gradients.ts
@@ -14,11 +14,25 @@
 import { varsThemeBase } from "../themes";
 import { createLinearGradient } from "./gradientBuilder";
 
-const interactiveVars = varsThemeBase.colors.aiGenerative.interactive;
+const { aiGradient } = varsThemeBase.colors;
 
-export const aiGradientInteractive = createLinearGradient(
-  "90deg",
-  `${interactiveVars.start} 5%`,
-  `${interactiveVars.middle} 40%`,
-  `${interactiveVars.end} 95%`
+export const aiGenerativeInteractive = createLinearGradient(
+  "50deg",
+  `${aiGradient.blueHigh} 28.22%`,
+  `${aiGradient.purpleHigh} 49.27%`,
+  `${aiGradient.pinkHigh} 83.12%`
+);
+
+export const aiGenerativeHover = createLinearGradient(
+  "50deg",
+  `${aiGradient.pinkHigh} 28.22%`,
+  `${aiGradient.purpleHigh} 49.27%`,
+  `${aiGradient.blueHigh} 83.12%`
+);
+
+export const aiGenerativeDisabled = createLinearGradient(
+  "50deg",
+  `${aiGradient.blueLight} 28.22%`,
+  `${aiGradient.purpleLight} 49.27%`,
+  `${aiGradient.pinkLight} 83.12%`
 );