+++ packages/core/styles/src/packages/atomic/box/nimbus-box.css.ts
@@ -34,13 +34,14 @@ import { mediaQueries, varsThemeBase } from "../../../themes";
 import { createBorderGradient, gradients } from "../../../gradients";
 
 const [, aiInteractiveBorderBg] = createBorderGradient(
-  gradients.aiGradientInteractive,
+  gradients.aiGenerativeInteractive,
   varsThemeBase.colors.neutral.background
 );
 
 const boxBackgroundProperties = {
   "ai-generative-interactive-border": aiInteractiveBorderBg,
-  "ai-generative-interactive": gradients.aiGradientInteractive,
+  "ai-generative-interactive": gradients.aiGenerativeInteractive,
+
   transparent: "transparent",
 };
 
@@ -77,6 +78,9 @@ const boxBackgroundColorProperties = {
   "neutral-surfaceDisabled":
     backgroundColorProperties["neutral-surfaceDisabled"],
   "neutral-textLow": backgroundColorProperties["neutral-textLow"],
+  "ai-generative-surface": backgroundColorProperties["ai-generativeSurface"],
+  "ai-generative-surfaceHighlight":
+    backgroundColorProperties["ai-generativeSurfaceHighlight"],
   transparent: "transparent",
 };