+++ packages/core/styles/src/packages/atomic/link/nimbus-link.css.ts
@@ -47,6 +47,9 @@ export const appearance = styleVariants({
       ":active": {
         color: varsThemeBase.colors.primary.interactivePressed,
       },
+      ":disabled": {
+        color: varsThemeBase.colors.neutral.textDisabled,
+      },
     },
   ],
   danger: [
@@ -58,6 +61,9 @@ export const appearance = styleVariants({
       },
       ":active": {
         color: varsThemeBase.colors.danger.interactivePressed,
+      },
+      ":disabled": {
+        color: varsThemeBase.colors.neutral.textDisabled,
       },
     },
   ],
@@ -71,6 +77,9 @@ export const appearance = styleVariants({
       ":active": {
         color: varsThemeBase.colors.neutral.textHigh,
       },
+      ":disabled": {
+        color: varsThemeBase.colors.neutral.textDisabled,
+      },
     },
   ],
   "neutral-background": [
@@ -82,6 +91,9 @@ export const appearance = styleVariants({
       },
       ":active": {
         color: varsThemeBase.colors.neutral.surfaceDisabled,
+      },
+      ":disabled": {
+        color: varsThemeBase.colors.neutral.textDisabled,
       },
     },
   ],