+++ packages/react/src/atomic/IconButton/src/iconButton.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { TiendanubeIcon } from "@nimbus-ds/icons";
+import { GenerativeStarsIcon, TiendanubeIcon } from "@nimbus-ds/icons";
 import { iconButton } from "@nimbus-ds/styles";
 import { argTypesConvert } from ".storybook/utils";
 import { IconButton } from "./IconButton";
@@ -13,6 +13,7 @@ const meta: Meta<typeof IconButton> = {
   argTypes: {
     source: { control: { disable: true } },
     ...argTypesConvert(iconButton.properties),
+    disabled: { control: "boolean" },
   },
   tags: ["autodocs"],
 };
@@ -38,7 +39,28 @@ export const primaryColor: Story = {
 export const aiGenerative: Story = {
   args: {
     size: "2.75rem",
+    color: "neutral-surface",
+    appearance: "ai-generative",
+    source: <GenerativeStarsIcon size="small" />,
+  },
+};
+
+export const aiGenerativeDisabled: Story = {
+  args: {
+    size: "2.75rem",
+    color: "neutral-textDisabled",
+    appearance: "ai-generative",
+    source: <GenerativeStarsIcon size="small" />,
+    disabled: true,
+  },
+};
+
+export const aiGenerativeSurface: Story = {
+  args: {
+    size: "2.75rem",
+    borderColor: "transparent",
     color: "ai-generative",
-    source: <TiendanubeIcon size="small" />,
+    backgroundColor: "ai-generativeSurface",
+    source: <GenerativeStarsIcon size="small" />,
   },
 };