+++ packages/react/src/atomic/Box/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 A low-level utility component that accepts styled system props to enable custom theme-aware styling
 
+## 2025-09-17 `4.3.1`
+
+#### 🐛 Bug fixes
+
+- Fixed `ai-generative-interactive` and `ai-generative-interactive-border` options to match Nimbus AI gradient style. ([#352](https://github.com/TiendaNube/nimbus-design-system/pull/352) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-09-16 `4.3.0`
 
 #### 🎉 New features